requirements.txt
Jinja2>=3.1.2
PyYAML>=6.0
Markdown>=3.4.1
python-dateutil>=2.8.2
python-frontmatter>=1.0.0
rcssmin>=1.1.1
rjsmin>=1.2.1
Pygments>=2.15.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
libsass>=0.23.0

